Moved some test functions to testutils
